Extract auxiliary-session tracking into a mixin (WS5 Seam 5)
Move the auxiliary / subagent session lifecycle tracking out of LCMEngine into
an AuxiliarySessionMixin in the new aux_session.py, and mix it in
(`class LCMEngine(ContextEngine, AuxiliarySessionMixin)`).

Moved verbatim (27 methods): the per-thread auxiliary stack, active/lineage
session-id sets, register/deactivate/handoff, the process-local generation
tokens, and the host-coupled caller-frame + state.db ancestor detection. They
keep running bound to the engine instance, so the auxiliary state (initialised
in __init__, cleared in _reset_profile_runtime_state, guarded by
_auxiliary_session_lock) and the callbacks (_end_host_fallback_compressor_for_session,
_state_db_path) stay on the engine and resolve via self.

The bypass-lineage family (which shares _auxiliary_session_lock but is a separate
concern) and the active-engine registry stay on the engine. engine.py shrinks by
~695 lines and drops its now-unused `inspect` import (moved to aux_session.py).
No call sites and no tests change; the ~117 white-box test references that call
these methods on an LCMEngine instance resolve via the mixin.

Behaviour-preserving refactor (WS5 engine decomposition). Rollback: git revert
this commit re-inlines the methods; no data or state migration involved.

Bases order: the mixin precedes ContextEngine so the LCM decomposition mixins
resolve consistently across seams (required for CompactionMixin, which overrides
the compress/should_compress protocol methods; harmless for private-only mixins).
